feat(factory-portal): Introduce FactoryLayout component for shared layout in protected routes
