+++ apps/factory-portal/src/App.jsx
@@ -8,6 +8,7 @@ import Inventory from './pages/Inventory';
 import InventoryDetail from './pages/InventoryDetail';
 import Orders from './pages/Orders';
 import OrderDetail from './pages/OrderDetail';
+import FactoryLayout from './components/FactoryLayout';
 
 // Create a theme with factory/admin colors
 const theme = createTheme({
@@ -31,19 +32,23 @@ function App() {
             {/* Public routes */}
             <Route path="/login" element={<Login />} />
 
-            {/* Protected routes for both admin and subadmin */}
+            {/* Protected routes with shared layout */}
             <Route element={<ProtectedRoute />}>
-              <Route path="/dashboard" element={<Dashboard />} />
-              <Route path="/inventory" element={<Inventory />} />
-              <Route path="/inventory/:id" element={<InventoryDetail />} />
-              <Route path="/orders" element={<Orders />} />
-              <Route path="/orders/:id" element={<OrderDetail />} />
-              {/* Add more protected routes here that both roles can access */}
+              <Route  element={<FactoryLayout />}>
+                <Route path="/dashboard" element={<Dashboard />} />
+                <Route path="/inventory" element={<Inventory />} />
+                <Route path="/inventory/:id" element={<InventoryDetail />} />
+                <Route path="/orders" element={<Orders />} />
+                <Route path="/orders/:id" element={<OrderDetail />} />
+                {/* Add more protected routes here that both roles can access */}
+              </Route>
             </Route>
 
             {/* Admin-only routes */}
             <Route element={<AdminRoute />}>
-              {/* Add admin-only routes here */}
+              <Route element={<FactoryLayout />}>
+                {/* Add admin-only routes here */}
+              </Route>
             </Route>
 
             {/* Redirect to login if no route matches */}
