+++ apps/factory-portal/src/components/FactoryLayout.jsx
@@ -1,0 +1,176 @@
+import { useState } from 'react';
+import { Link, useNavigate, Outlet } from 'react-router-dom';
+import { useAuth } from '../contexts/AuthContext';
+import {
+    Box,
+    AppBar,
+    Toolbar,
+    IconButton,
+    Typography,
+    Drawer,
+    List,
+    ListItem,
+    ListItemIcon,
+    ListItemText,
+    Divider,
+    Badge,
+    Menu,
+    MenuItem
+} from '@mui/material';
+import {
+    AccountCircle,
+    Menu as MenuIcon,
+    Dashboard as DashboardIcon,
+    Inventory as InventoryIcon,
+    ShoppingCart as ShoppingCartIcon,
+    People as PeopleIcon,
+    Settings as SettingsIcon,
+    Logout as LogoutIcon
+} from '@mui/icons-material';
+
+const drawerWidth = 240;
+
+const FactoryLayout = () => {
+    const { admin, logout, isAdmin } = useAuth();
+    const navigate = useNavigate();
+    const [anchorEl, setAnchorEl] = useState(null);
+    const [drawerOpen, setDrawerOpen] = useState(true);
+
+    const handleMenu = (event) => {
+        setAnchorEl(event.currentTarget);
+    };
+
+    const handleClose = () => {
+        setAnchorEl(null);
+    };
+
+    const handleLogout = () => {
+        logout();
+        navigate('/login');
+    };
+
+    return (
+        <Box sx={{ display: 'flex' }}>
+            <AppBar position="fixed" sx={{ zIndex: (theme) => theme.zIndex.drawer + 1 }}>
+                <Toolbar>
+                    <IconButton
+                        color="inherit"
+                        aria-label="open drawer"
+                        edge="start"
+                        onClick={() => setDrawerOpen(!drawerOpen)}
+                        sx={{ mr: 2 }}
+                    >
+                        <MenuIcon />
+                    </IconButton>
+                    <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+                        LogiTrack Factory Portal
+                    </Typography>
+                    <div>
+                        <IconButton
+                            size="large"
+                            aria-label="account of current user"
+                            aria-controls="menu-appbar"
+                            aria-haspopup="true"
+                            onClick={handleMenu}
+                            color="inherit"
+                        >
+                            <Badge
+                                color="secondary"
+                                variant="dot"
+                                invisible={!(admin?.role === 'admin')}
+                            >
+                                <AccountCircle />
+                            </Badge>
+                        </IconButton>
+                        <Menu
+                            id="menu-appbar"
+                            anchorEl={anchorEl}
+                            anchorOrigin={{
+                                vertical: 'top',
+                                horizontal: 'right',
+                            }}
+                            keepMounted
+                            transformOrigin={{
+                                vertical: 'top',
+                                horizontal: 'right',
+                            }}
+                            open={Boolean(anchorEl)}
+                            onClose={handleClose}
+                        >
+                            <MenuItem onClick={handleClose}>Profile</MenuItem>
+                            <MenuItem onClick={handleClose}>Settings</MenuItem>
+                            <MenuItem onClick={handleLogout}>Logout</MenuItem>
+                        </Menu>
+                    </div>
+                </Toolbar>
+            </AppBar>
+
+            <Drawer
+                variant="persistent"
+                open={drawerOpen}
+                sx={{
+                    width: drawerWidth,
+                    flexShrink: 0,
+                    [`& .MuiDrawer-paper`]: { width: drawerWidth, boxSizing: 'border-box' },
+                }}
+            >
+                <Toolbar />
+                <Box sx={{ overflow: 'auto' }}>
+                    <List>
+                        <ListItem button component={Link} to="/dashboard">
+                            <ListItemIcon>
+                                <DashboardIcon />
+                            </ListItemIcon>
+                            <ListItemText primary="Dashboard" />
+                        </ListItem>
+                        <ListItem button component={Link} to="/inventory">
+                            <ListItemIcon>
+                                <InventoryIcon />
+                            </ListItemIcon>
+                            <ListItemText primary="Inventory" />
+                        </ListItem>
+                        <ListItem button component={Link} to="/orders">
+                            <ListItemIcon>
+                                <ShoppingCartIcon />
+                            </ListItemIcon>
+                            <ListItemText primary="Orders" />
+                        </ListItem>
+                        <ListItem button>
+                            <ListItemIcon>
+                                <PeopleIcon />
+                            </ListItemIcon>
+                            <ListItemText primary="Users" />
+                        </ListItem>
+                    </List>
+                    <Divider />
+                    {isAdmin && isAdmin() && (
+                        <List>
+                            <ListItem button>
+                                <ListItemIcon>
+                                    <SettingsIcon />
+                                </ListItemIcon>
+                                <ListItemText primary="Admin Settings" />
+                            </ListItem>
+                        </List>
+                    )}
+                    <Divider />
+                    <List>
+                        <ListItem button onClick={handleLogout}>
+                            <ListItemIcon>
+                                <LogoutIcon />
+                            </ListItemIcon>
+                            <ListItemText primary="Logout" />
+                        </ListItem>
+                    </List>
+                </Box>
+            </Drawer>
+
+            <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
+                <Toolbar />
+                <Outlet />
+            </Box>
+        </Box>
+    );
+};
+
+export default FactoryLayout;
